+++ lib/pack_app/navigator/app_navigator_popper.dart
@@ -3,10 +3,12 @@ part of pack_app;
 class AppNavigatorPopper extends AppNavigatorDataProvider {
   AppNavigatorPopper({
     required final RouteState routeState,
+    required final ScreenLayout screenLayout,
     required final BuildContext context,
   }) : super(
           routeState: routeState,
           context: context,
+          screenLayout: screenLayout,
         );
 
   Future<bool> handleWillPop() async {
@@ -18,7 +20,17 @@ class AppNavigatorPopper extends AppNavigatorDataProvider {
       case AppRoutesName.note:
       case AppRoutesName.createIdea:
       case AppRoutesName.settings:
+      case AppRoutesName.generalSettings:
         navigatorController.goHome();
+        break;
+      case AppRoutesName.subscription:
+      case AppRoutesName.changelog:
+      case AppRoutesName.profile:
+        if (navigatorController.screenLayout.small) {
+          navigatorController.goSettings();
+        } else {
+          navigatorController.goHome();
+        }
         break;
       case AppRoutesName.home:
     }
@@ -33,19 +45,24 @@ class AppNavigatorPopper extends AppNavigatorDataProvider {
     /// ! here will go selected pages logic.
     final maybePage = route.settings;
     if (maybePage is Page) {
-      if (maybePage.key == _ValueKeys._createIdea) {
+      if (maybePage.key == NavigatorValueKeys._createIdea) {
         navigatorController.goHome();
-      } else if (maybePage.key == _ValueKeys._ideasIdea) {
+      } else if (maybePage.key == NavigatorValueKeys._ideasIdea) {
         navigatorController.goHome();
-      } else if (maybePage.key == _ValueKeys._ideasIdeaAnswer) {
+      } else if (maybePage.key == NavigatorValueKeys._ideasIdeaAnswer) {
         final arr = maybePage.name?.split('/') ?? [];
         if (arr.length == 4) {
           navigatorController.goIdeaScreen(ideaId: arr[4]);
         } else {
           navigatorController.goHome();
         }
-      } else if (maybePage.key == _ValueKeys._settings) {
-        navigatorController.goHome();
+      } else if (maybePage.name?.contains(AppRoutesName.settings) == true) {
+        if (maybePage.name == AppRoutesName.settings ||
+            navigatorController.screenLayout.notSmall) {
+          navigatorController.goHome();
+        } else {
+          navigatorController.goSettings();
+        }
       }
     }