+++ lib/pack_app/screens/app/app_state_provider.dart
@@ -6,11 +6,14 @@ class AppStateProvider extends StatelessWidget {
     final Key? key,
   }) : super(key: key);
   final WidgetBuilder builder;
-  SettingsController get _settings => GlobalStateNotifiers.settings;
+  GeneralSettingsController get _settings => GlobalStateNotifiers.settings;
   @override
   Widget build(final BuildContext context) {
-    return MultiProvider(
+    final child = MultiProvider(
       providers: [
+        /// Keep _settings is global is important as it will not lose all
+        /// changes during global rebuild
+        ChangeNotifierProvider(create: (final _) => _settings),
         ChangeNotifierProvider(create: createEmojiProvider),
         ChangeNotifierProvider(create: createLastUsedEmojisProvider),
         ChangeNotifierProvider(create: createSpecialEmojisProvider),
@@ -20,22 +23,34 @@ class AppStateProvider extends StatelessWidget {
         ChangeNotifierProvider(create: createIdeaProjectQuestionsProvider),
         ChangeNotifierProvider(create: createNoteProjectsProvider),
         ChangeNotifierProvider(create: createNotificationController),
+        ChangeNotifierProvider(create: createPaymentsController),
       ],
       child: Portal(
-        child: SettingsStateScope(
-          notifier: _settings,
-          child: Builder(
-            builder: (final context) {
-              return StateLoader(
-                initializer: GlobalStateInitializer(
-                  settings: _settings,
-                ),
-                loader: const AppLoadingScreen(),
-                child: builder(context),
-              );
-            },
-          ),
+        child: Builder(
+          builder: (final context) {
+            return StateLoader(
+              initializer: GlobalStateInitializer(
+                settings: _settings,
+              ),
+              loader: const AppLoadingScreen(),
+              child: builder(context),
+            );
+          },
         ),
+      ),
+    );
+    if (isNativeDesktop) {
+      return child;
+    }
+
+    return Directionality(
+      // TODO(arenukvern): replace with default device textDirection
+      textDirection: TextDirection.ltr,
+      child: Stack(
+        children: [
+          Container(color: AppColors.black),
+          child,
+        ],
       ),
     );
   }