+++ lib/pack_auth/pack_auth.dart
@@ -1,0 +1,1 @@
+library pack_auth;